fix: make windows compatible
