+++ prettytable/termwidth_unix.go
@@ -1,0 +1,32 @@
+// Copyright 2019 Yunion
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build !windows
+
+package prettytable
+
+import (
+	"os"
+
+	"golang.org/x/sys/unix"
+)
+
+func termWidth() (int, error) {
+	fd := int(os.Stdout.Fd())
+	wsz, err := unix.IoctlGetWinsize(fd, unix.TIOCGWINSZ)
+	if err != nil {
+		return -1, err
+	}
+	return int(wsz.Col), nil
+}
